MDL-86301 core: Create autoloadable externallib_testcase
